+++ config/postcss.config.js
@@ -1,0 +1,11 @@
+/* globals module */
+
+module.exports = () => ( {
+	syntax: `postcss-scss`,
+	map: false,
+	plugins: {
+		'css-declaration-sorter': {
+			order: `smacss`
+		}
+	}
+} );