+++ includes/classes/abstracts/template.php
@@ -3,6 +3,16 @@
 if( !defined( 'ABSPATH' ) ) exit; // Exit if accessed directly
 
 abstract class BFG_Abstract_Page_Template {
+	/**
+	 * Defines the number of display_image() images that shouldn't be lazy loaded on the page
+	 * Set to 0 to lazy load all images
+	 *
+	 * @since 20220126
+	 */
+	protected $omit_threshold = 1;
+
+	protected $image_count = 0;
+
 	private function get_value($key, $parent) {
 
 		if( $parent === false ) {
@@ -60,6 +70,21 @@ abstract class BFG_Abstract_Page_Template {
 
 	}
 
+	private function get_lazy_loading_attr() {
+
+		if( !wp_lazy_loading_enabled( 'img', 'wp_get_attachment_image' ) )
+			return 'eager';
+
+		if( $this->omit_threshold === 0 )
+			return 'lazy';
+
+		if( $this->image_count < $this->omit_threshold )
+			return 'eager';
+
+		return 'lazy';
+
+	}
+
 	protected function display_image($key, $width, $height, $parent = false) {
 
 		$image_id = $this->get_value($key, $parent);
@@ -72,7 +97,7 @@ abstract class BFG_Abstract_Page_Template {
 				$alt = get_post_meta( $image_id, '_wp_attachment_image_alt', true );
 
 				?>
-				<img src="<?php echo wp_get_attachment_url( $image_id ); ?>" alt="<?php echo esc_attr($alt); ?>" width="<?php echo $width; ?>" height="<?php echo $height; ?>" <?php echo wp_lazy_loading_enabled( 'img', 'wp_get_attachment_image' ) ? 'loading="lazy"' : ''; ?>>
+				<img src="<?php echo wp_get_attachment_url( $image_id ); ?>" alt="<?php echo esc_attr($alt); ?>" width="<?php echo $width; ?>" height="<?php echo $height; ?>" loading="<?php echo $this->get_lazy_loading_attr(); ?>">
 				<?php
 				break;
 			default:
@@ -81,14 +106,26 @@ abstract class BFG_Abstract_Page_Template {
 						$image_id,
 						array(
 							array($width, $height, true),
+						),
+						array(
+							'loading' => $this->get_lazy_loading_attr(),
 						)
 					);
 				} else {
-					echo wp_get_attachment_image( $image_id, array($width, $height) );
+					echo wp_get_attachment_image(
+						$image_id,
+						array($width, $height),
+						false,
+						array(
+							'loading' => $this->get_lazy_loading_attr(),
+						)
+					);
 				}
 
 				break;
 		}
+
+		++$this->image_count;
 
 	}
 }